Okay, we're making excellent progress!
home.html is now building successfully! This is a great step.
The error for investment.html has changed to 'loop' is undefined. This is a Jinja2 error indicating that loop.index or a similar loop variable is being used outside of a {% for %} block, or the loop itself has an issue.